+++ addOns/spiderAjax/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Changed
+- Updated automation framework documentation and templates for `spiderAjax` job to reflect changes to the default value of numberOfBrowsers parameter
 - Fields with default or missing values are omitted for the `spiderAjax` job in saved Automation Framework plans.
 
 ## [23.21.0] - 2024-09-02