+++ addOns/spiderAjax/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [23.22.0] - 2025-01-10
 ### Added
 - Option to enable browser extensions added by other add-ons, previously they were always enabled but now the default is false.
 
@@ -332,6 +332,7 @@ By default it allows files with extension `.js` and `.css`.
 
 
 
+[23.22.0]: https://github.com/zaproxy/zap-extensions/releases/spiderAjax-v23.22.0
 [23.21.0]: https://github.com/zaproxy/zap-extensions/releases/spiderAjax-v23.21.0
 [23.20.0]: https://github.com/zaproxy/zap-extensions/releases/spiderAjax-v23.20.0
 [23.19.0]: https://github.com/zaproxy/zap-extensions/releases/spiderAjax-v23.19.0